Remove protocol module and refactor jsonrpc files
