+++ trin-core/src/jsonrpc/mod.rs
@@ -1,0 +1,3 @@
+pub mod handlers;
+pub mod service;
+pub mod types;
