+++ trin-core/src/jsonrpc/service.rs
@@ -1,9 +1,9 @@
+use super::handlers::{PortalEndpoint, PortalEndpointKind};
+use super::types::JsonRequest;
 use crate::cli::TrinConfig;
-use crate::portalnet::protocol::{PortalEndpoint, PortalEndpointKind};
 use log::debug;
 use reqwest::blocking as reqwest;
-use serde::{Deserialize, Serialize};
-use serde_json::{json, Map, Value};
+use serde_json::{json, Value};
 use std::fs;
 use std::io::{self, Read, Write};
 use std::net::{TcpListener, TcpStream};
@@ -16,50 +16,7 @@ use std::{panic, process};
 use threadpool::ThreadPool;
 use tokio::sync::mpsc;
 use tokio::sync::mpsc::UnboundedSender;
-use validator::{Validate, ValidationError};
-
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-#[serde(untagged)]
-pub enum Params {
-    /// No parameters
-    None,
-    /// Array of values
-    Array(Vec<Value>),
-    /// Map of values
-    Map(Map<String, Value>),
-}
-
-impl From<Params> for Value {
-    fn from(params: Params) -> Value {
-        match params {
-            Params::Array(vec) => Value::Array(vec),
-            Params::Map(map) => Value::Object(map),
-            Params::None => Value::Null,
-        }
-    }
-}
-
-#[derive(Debug, Deserialize, Serialize, Validate)]
-struct JsonRequest {
-    #[validate(custom = "validate_jsonrpc_version")]
-    pub jsonrpc: String,
-    #[serde(default = "default_params")]
-    pub params: Params,
-    pub method: String,
-    pub id: u32,
-}
-
-fn default_params() -> Params {
-    Params::None
-}
-
-fn validate_jsonrpc_version(jsonrpc: &str) -> Result<(), ValidationError> {
-    if jsonrpc != "2.0" {
-        return Err(ValidationError::new("Unsupported jsonrpc version"));
-    }
-    Ok(())
-}
+use validator::Validate;
 
 lazy_static! {
     static ref IPC_PATH: Mutex<String> = Mutex::new(String::new());
@@ -373,55 +330,4 @@ fn proxy_to_url(request: String, url: &str) -> io::Result<Vec<u8>> {
 
 fn get_infura_url(infura_project_id: &str) -> String {
     return format!("https://mainnet.infura.io:443/v3/{}", infura_project_id);
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use rstest::rstest;
-    use validator::ValidationErrors;
-
-    #[test]
-    fn test_json_validator_accepts_valid_json() {
-        let request = JsonRequest {
-            jsonrpc: "2.0".to_string(),
-            id: 1,
-            params: Params::None,
-            method: "eth_blockNumber".to_string(),
-        };
-        assert_eq!(request.validate(), Ok(()));
-    }
-
-    #[test]
-    fn test_json_validator_with_invalid_jsonrpc_field() {
-        let request = JsonRequest {
-            jsonrpc: "1.0".to_string(),
-            id: 1,
-            params: Params::None,
-            method: "eth_blockNumber".to_string(),
-        };
-        let errors = request.validate();
-        assert!(ValidationErrors::has_error(&errors, "jsonrpc"));
-    }
-
-    fn expected_map() -> Map<String, Value> {
-        let mut expected_map = serde_json::Map::new();
-        expected_map.insert("key".to_string(), Value::String("value".to_string()));
-        expected_map
-    }
-
-    #[rstest]
-    #[case("[null]", Params::Array(vec![Value::Null]))]
-    #[case("[true]", Params::Array(vec![Value::Bool(true)]))]
-    #[case("[-1]", Params::Array(vec![Value::from(-1)]))]
-    #[case("[4]", Params::Array(vec![Value::from(4)]))]
-    #[case("[2.3]", Params::Array(vec![Value::from(2.3)]))]
-    #[case("[\"hello\"]", Params::Array(vec![Value::String("hello".to_string())]))]
-    #[case("[[0]]", Params::Array(vec![Value::Array(vec![Value::from(0)])]))]
-    #[case("[[]]", Params::Array(vec![Value::Array(vec![])]))]
-    #[case("[{\"key\": \"value\"}]", Params::Array(vec![Value::Object(expected_map())]))]
-    fn request_params_deserialization(#[case] input: &str, #[case] expected: Params) {
-        let deserialized: Params = serde_json::from_str(input).unwrap();
-        assert_eq!(deserialized, expected);
-    }
 }
