+++ trin-core/src/portalnet/mod.rs
@@ -7,7 +7,6 @@ use uint::construct_uint;
 pub mod discovery;
 pub mod events;
 pub mod overlay;
-pub mod protocol;
 pub mod types;
 pub mod utp;
